+++ docsrc/source/changelog.rst
@@ -4,6 +4,12 @@ Changelog
 Here you will find a list of changes for each package update related to the Relevance AI
 Python library.
 
+v0.33.3
+--------
+
+- Fix delete dataset for international gateways
+- Add low-touch way to label with a given model
+
 v0.33.2
 --------